Refactor parts of ApiService to ExerciseService
