+++ frontend/src/app/pages/exercises/exercise/shared/alarm-group-overview/alarm-group-item/alarm-group-item.component.ts
@@ -2,7 +2,7 @@ import { Component, Input } from '@angular/core';
 import { Store } from '@ngrx/store';
 import type { UUID } from 'digital-fuesim-manv-shared';
 import { AlarmGroup, AlarmGroupVehicle } from 'digital-fuesim-manv-shared';
-import { ApiService } from 'src/app/core/api.service';
+import { ExerciseService } from 'src/app/core/exercise.service';
 import type { AppState } from 'src/app/state/app.state';
 import {
     getSelectVehicleTemplate,
@@ -19,7 +19,7 @@ export class AlarmGroupItemComponent {
     @Input() alarmGroup!: AlarmGroup;
 
     constructor(
-        private readonly apiService: ApiService,
+        private readonly exerciseService: ExerciseService,
         private readonly store: Store<AppState>
     ) {}
 
@@ -28,7 +28,7 @@ export class AlarmGroupItemComponent {
     );
 
     public renameAlarmGroup(name: string) {
-        this.apiService.proposeAction(
+        this.exerciseService.proposeAction(
             {
                 type: '[AlarmGroup] Rename AlarmGroup',
                 alarmGroupId: this.alarmGroup.id,
@@ -39,14 +39,14 @@ export class AlarmGroupItemComponent {
     }
 
     public removeAlarmGroup() {
-        this.apiService.proposeAction({
+        this.exerciseService.proposeAction({
             type: '[AlarmGroup] Remove AlarmGroup',
             alarmGroupId: this.alarmGroup.id,
         });
     }
 
     public removeVehicleTemplate(alarmGroupVehicleId: UUID) {
-        this.apiService.proposeAction({
+        this.exerciseService.proposeAction({
             type: '[AlarmGroup] Remove AlarmGroupVehicle',
             alarmGroupId: this.alarmGroup.id,
             alarmGroupVehicleId,
@@ -64,7 +64,7 @@ export class AlarmGroupItemComponent {
         if (name === null) {
             return;
         }
-        this.apiService.proposeAction(
+        this.exerciseService.proposeAction(
             {
                 type: '[AlarmGroup] Edit AlarmGroupVehicle',
                 alarmGroupId: this.alarmGroup.id,
@@ -81,7 +81,7 @@ export class AlarmGroupItemComponent {
             getSelectVehicleTemplate(vehicleTemplateId),
             this.store
         )!;
-        this.apiService.proposeAction({
+        this.exerciseService.proposeAction({
             type: '[AlarmGroup] Add AlarmGroupVehicle',
             alarmGroupId: this.alarmGroup.id,
             alarmGroupVehicle: AlarmGroupVehicle.create(
