+++ frontend/src/app/pages/exercises/exercise/shared/client-overview/client-overview-table/client-overview-table.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
 import type { UUID } from 'digital-fuesim-manv-shared';
-import { ApiService } from 'src/app/core/api.service';
+import { ExerciseService } from 'src/app/core/exercise.service';
 import type { AppState } from 'src/app/state/app.state';
 import {
     selectClients,
@@ -19,14 +19,14 @@ export class ClientOverviewTableComponent {
 
     constructor(
         private readonly store: Store<AppState>,
-        private readonly apiService: ApiService
+        private readonly exerciseService: ExerciseService
     ) {}
 
     public async restrictToViewport(
         clientId: UUID,
         viewportId: UUID | undefined
     ) {
-        this.apiService.proposeAction({
+        this.exerciseService.proposeAction({
             type: '[Client] Restrict to viewport',
             clientId,
             viewportId,
@@ -37,7 +37,7 @@ export class ClientOverviewTableComponent {
         clientId: UUID,
         shouldBeInWaitingRoom: boolean
     ) {
-        this.apiService.proposeAction({
+        this.exerciseService.proposeAction({
             type: '[Client] Set waitingroom',
             clientId,
             shouldBeInWaitingRoom,
