+++ frontend/src/app/pages/exercises/exercise/shared/core/drag-element.service.ts
@@ -1,20 +1,20 @@
 import { Injectable } from '@angular/core';
-import { ApiService } from 'src/app/core/api.service';
-import type OlMap from 'ol/Map';
 import type {
     ImageProperties,
     MapImageTemplate,
     VehicleTemplate,
 } from 'digital-fuesim-manv-shared';
 import {
-    Viewport,
-    TransferPoint,
-    normalZoom,
     createVehicleParameters,
-    PatientTemplate,
     MapImage,
+    normalZoom,
+    PatientTemplate,
+    TransferPoint,
+    Viewport,
 } from 'digital-fuesim-manv-shared';
 import type { PatientCategory } from 'digital-fuesim-manv-shared/dist/models/patient-category';
+import type OlMap from 'ol/Map';
+import { ExerciseService } from 'src/app/core/exercise.service';
 
 @Injectable({
     providedIn: 'root',
@@ -25,7 +25,7 @@ import type { PatientCategory } from 'digital-fuesim-manv-shared/dist/models/pat
 export class DragElementService {
     private olMap?: OlMap;
 
-    constructor(private readonly apiService: ApiService) {}
+    constructor(private readonly exerciseService: ExerciseService) {}
 
     public registerMap(map: OlMap) {
         this.olMap = map;
@@ -128,7 +128,7 @@ export class DragElementService {
         // create the element
         switch (this.transferringTemplate.type) {
             case 'vehicle':
-                this.apiService.proposeAction(
+                this.exerciseService.proposeAction(
                     {
                         type: '[Vehicle] Add vehicle',
                         ...createVehicleParameters(
@@ -151,7 +151,7 @@ export class DragElementService {
                         ]!,
                         this.transferringTemplate.template.name
                     );
-                    this.apiService.proposeAction(
+                    this.exerciseService.proposeAction(
                         {
                             type: '[Patient] Add patient',
                             patient: {
@@ -167,7 +167,7 @@ export class DragElementService {
                 // This ratio has been determined by trial and error
                 const height = Viewport.image.height / 23.5;
                 const width = height * Viewport.image.aspectRatio;
-                this.apiService.proposeAction(
+                this.exerciseService.proposeAction(
                     {
                         type: '[Viewport] Add viewport',
                         viewport: Viewport.create(
@@ -189,14 +189,14 @@ export class DragElementService {
             case 'mapImage':
                 {
                     const template = this.transferringTemplate.template.image;
-                    this.apiService.proposeAction({
+                    this.exerciseService.proposeAction({
                         type: '[MapImage] Add MapImage',
                         mapImage: MapImage.create(position, template, false),
                     });
                 }
                 break;
             case 'transferPoint':
-                this.apiService.proposeAction(
+                this.exerciseService.proposeAction(
                     {
                         type: '[TransferPoint] Add TransferPoint',
                         transferPoint: TransferPoint.create(
