+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/exercise-map.component.ts
@@ -8,11 +8,11 @@ import {
 } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Subject, takeUntil } from 'rxjs';
-import { ApiService } from 'src/app/core/api.service';
+import { ExerciseService } from 'src/app/core/exercise.service';
 import type { AppState } from 'src/app/state/app.state';
 import {
-    selectRestrictedViewport,
     selectCurrentRole,
+    selectRestrictedViewport,
 } from 'src/app/state/application/selectors/shared.selectors';
 import { DragElementService } from '../core/drag-element.service';
 import { TransferLinesService } from '../core/transfer-lines.service';
@@ -43,7 +43,7 @@ export class ExerciseMapComponent implements AfterViewInit, OnDestroy {
     constructor(
         private readonly store: Store<AppState>,
         private readonly ngZone: NgZone,
-        private readonly apiService: ApiService,
+        private readonly exerciseService: ExerciseService,
         public readonly dragElementService: DragElementService,
         public readonly transferLinesService: TransferLinesService
     ) {}
@@ -53,7 +53,7 @@ export class ExerciseMapComponent implements AfterViewInit, OnDestroy {
         this.ngZone.runOutsideAngular(() => {
             this.olMapManager = new OlMapManager(
                 this.store,
-                this.apiService,
+                this.exerciseService,
                 this.openLayersContainer.nativeElement,
                 this.popoverContainer.nativeElement,
                 this.ngZone,
