+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/feature-managers/map-images-feature-manager.ts
@@ -5,10 +5,10 @@ import type Point from 'ol/geom/Point';
 import type VectorLayer from 'ol/layer/Vector';
 import type OlMap from 'ol/Map';
 import type VectorSource from 'ol/source/Vector';
-import type { ApiService } from 'src/app/core/api.service';
+import type { ExerciseService } from 'src/app/core/exercise.service';
 import type { AppState } from 'src/app/state/app.state';
-import { selectStateSnapshot } from 'src/app/state/get-state-snapshot';
 import { selectCurrentRole } from 'src/app/state/application/selectors/shared.selectors';
+import { selectStateSnapshot } from 'src/app/state/get-state-snapshot';
 import { MapImagePopupComponent } from '../shared/map-image-popup/map-image-popup.component';
 import { ImagePopupHelper } from '../utility/popup-helper';
 import { ImageStyleHelper } from '../utility/style-helper/image-style-helper';
@@ -24,14 +24,14 @@ export class MapImageFeatureManager extends ElementFeatureManager<MapImage> {
     constructor(
         olMap: OlMap,
         layer: VectorLayer<VectorSource<Point>>,
-        apiService: ApiService,
+        exerciseService: ExerciseService,
         private readonly store: Store<AppState>
     ) {
         super(
             olMap,
             layer,
             (targetPosition, mapImage) => {
-                apiService.proposeAction({
+                exerciseService.proposeAction({
                     type: '[MapImage] Move MapImage',
                     mapImageId: mapImage.id,
                     targetPosition,
