+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/feature-managers/transfer-point-feature-manager.ts
@@ -7,10 +7,10 @@ import type { TranslateEvent } from 'ol/interaction/Translate';
 import type VectorLayer from 'ol/layer/Vector';
 import type OlMap from 'ol/Map';
 import type VectorSource from 'ol/source/Vector';
-import type { ApiService } from 'src/app/core/api.service';
+import type { ExerciseService } from 'src/app/core/exercise.service';
 import type { AppState } from 'src/app/state/app.state';
-import { selectStateSnapshot } from 'src/app/state/get-state-snapshot';
 import { selectCurrentRole } from 'src/app/state/application/selectors/shared.selectors';
+import { selectStateSnapshot } from 'src/app/state/get-state-snapshot';
 import { ChooseTransferTargetPopupComponent } from '../shared/choose-transfer-target-popup/choose-transfer-target-popup.component';
 import { TransferPointPopupComponent } from '../shared/transfer-point-popup/transfer-point-popup.component';
 import { ImagePopupHelper } from '../utility/popup-helper';
@@ -26,13 +26,13 @@ export class TransferPointFeatureManager extends ElementFeatureManager<TransferP
         olMap: OlMap,
         layer: VectorLayer<VectorSource<Point>>,
         private readonly store: Store<AppState>,
-        private readonly apiService: ApiService
+        private readonly exerciseService: ExerciseService
     ) {
         super(
             olMap,
             layer,
             (targetPosition, transferPoint) => {
-                apiService.proposeAction({
+                exerciseService.proposeAction({
                     type: '[TransferPoint] Move TransferPoint',
                     transferPointId: transferPoint.id,
                     targetPosition,
@@ -102,7 +102,7 @@ export class TransferPointFeatureManager extends ElementFeatureManager<TransferP
                         targetType: 'hospital' | 'transferPoint'
                     ) => {
                         if (targetType === 'hospital') {
-                            this.apiService.proposeAction(
+                            this.exerciseService.proposeAction(
                                 {
                                     type: '[Hospital] Transport patient to hospital',
                                     hospitalId: targetId,
@@ -112,7 +112,7 @@ export class TransferPointFeatureManager extends ElementFeatureManager<TransferP
                             );
                             return;
                         }
-                        this.apiService.proposeAction(
+                        this.exerciseService.proposeAction(
                             {
                                 type: '[Transfer] Add to transfer',
                                 elementType: droppedElement.type,
