+++ frontend/src/app/pages/exercises/exercise/shared/transfer-overview/transfer-time-input/transfer-time-input.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Transfer, UUID } from 'digital-fuesim-manv-shared';
-import { ApiService } from 'src/app/core/api.service';
+import { ExerciseService } from 'src/app/core/exercise.service';
 import type { AppState } from 'src/app/state/app.state';
 import { selectCurrentTime } from 'src/app/state/application/selectors/exercise.selectors';
 
@@ -21,11 +21,11 @@ export class TransferTimeInputComponent {
 
     constructor(
         private readonly store: Store<AppState>,
-        private readonly apiService: ApiService
+        private readonly exerciseService: ExerciseService
     ) {}
 
     public addTransferTime(timeToAdd: number) {
-        this.apiService.proposeAction({
+        this.exerciseService.proposeAction({
             type: '[Transfer] Edit transfer',
             elementType: this.elementType,
             elementId: this.elementId,
@@ -34,7 +34,7 @@ export class TransferTimeInputComponent {
     }
 
     public togglePauseTransfer() {
-        this.apiService.proposeAction({
+        this.exerciseService.proposeAction({
             type: '[Transfer] Toggle pause transfer',
             elementType: this.elementType,
             elementId: this.elementId,
@@ -42,7 +42,7 @@ export class TransferTimeInputComponent {
     }
 
     public letElementArrive() {
-        this.apiService.proposeAction({
+        this.exerciseService.proposeAction({
             type: '[Transfer] Finish transfer',
             elementType: this.elementType,
             elementId: this.elementId,
